fix: Event index

+++ Source/Client/Game/Objects/Map.cs
@@ -1204,7 +1204,7 @@ namespace Client
                                         {
                                             {
                                                 ref var withBlock2 = ref Core.Type.MyMap.Event[i].Pages[x].CommandList[y].Commands[z];
-                                                buffer.WriteByte(withBlock2.Index);
+                                                buffer.WriteInt32(withBlock2.Index);
                                                 buffer.WriteString(withBlock2.Text1);
                                                 buffer.WriteString(withBlock2.Text2);
                                                 buffer.WriteString(withBlock2.Text3);
